requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-braces==1.14.0
django-crispy-forms==1.11.1
django-froala-editor==3.2.2
django-import-export==2.5.0
django-model-utils==4.1.1
django-ratings==0.3.7
django-star-ratings==0.9.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
# paytm==0.1.8
Pillow==8.1.0
# pycrypto==2.6.1
pydal==20200714.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
